Refactor TeamsRepositoryImpl to remove UploadManager dependency
- Decoupled `UploadManager` from `TeamsRepositoryImpl` by removing the dependency and moving the upload logic up to the ViewModel layer.
- `TeamViewModel`, `RequestsViewModel`, `TeamDetailFragment`, and `TeamResourcesFragment` now inject `UploadManager` directly and call `uploadManager.uploadTeams()`, `uploadTeamActivities()`, and `uploadResource()` after invoking `teamsSyncRepository.syncTeamActivities()`.
- Restored `UploadManager.kt` to its original state, avoiding the creation of the flawed `TeamsUploadRepository` abstraction which simply relocated the dependency inversion.
- Updated unit tests (`TeamViewModelTest`, `RequestsViewModelTest`, `TeamsRepositoryBulkInsertTransactionTest`, `TeamsRepositoryImplTest`) to properly mock and accommodate the modified constructors and logic flows.
