+++ app/src/test/java/org/ole/planet/myplanet/services/UploadManagerTest.kt
@@ -40,7 +40,6 @@ import org.ole.planet.myplanet.repository.SubmissionsRepository
 import org.ole.planet.myplanet.repository.TeamUploadData
 import org.ole.planet.myplanet.repository.TeamsRepository
 import org.ole.planet.myplanet.repository.TeamsSyncRepository
-import org.ole.planet.myplanet.repository.TeamsUploadRepository
 import org.ole.planet.myplanet.repository.UploadRepository
 import org.ole.planet.myplanet.repository.UserRepository
 import org.ole.planet.myplanet.repository.VoicesRepository
@@ -60,18 +59,51 @@ class UploadManagerTest {
 
     @Test
     fun `uploadTeams handles bulk success`() = testScope.runTest {
-        coEvery { teamsUploadRepository.uploadTeams() } returns Unit
-        teamsUploadRepository.uploadTeams()
+
+        val mockTeam1 = TeamUploadData("team1", JsonObject(), false, null)
+        val mockTeam2 = TeamUploadData("team2", JsonObject(), false, null)
+        val mockTeam3 = TeamUploadData("team3", JsonObject(), true, null)
+        val mockRepo = io.mockk.mockk<TeamsSyncRepository>(relaxed = true)
+        every { teamsSyncRepository.get() } returns mockRepo
+        coEvery { mockRepo.getTeamsForUpload() } returns listOf(mockTeam1, mockTeam2, mockTeam3)
+
+        val bulkResponse = com.google.gson.JsonArray().apply {
+            add(JsonObject().apply { addProperty("id", "team1"); addProperty("rev", "rev1") })
+            add(JsonObject().apply { addProperty("id", "team2"); addProperty("error", "conflict") })
+            add(JsonObject().apply { addProperty("id", "team3"); addProperty("rev", "rev3") })
+        }
+        coEvery { uploadRepository.postUploadArray(any(), any()) } returns retrofit2.Response.success(bulkResponse)
+
+        coEvery { retryQueue.queueFailedOperation(any(), any(), any(), any(), any(), any(), any()) } returns Unit
+        coEvery { mockRepo.markTeamsUploaded(any()) } returns Unit
+        coEvery { mockRepo.deleteLocalTeamRecords(any()) } returns Unit
+
+        uploadManager.uploadTeams()
         advanceUntilIdle()
-        coVerify { teamsUploadRepository.uploadTeams() }
+
+        coVerify(exactly = 1) { uploadRepository.postUploadArray("http://mock.url/teams/_bulk_docs", any()) }
+        coVerify(exactly = 1) { mockRepo.markTeamsUploaded(mapOf("team1" to "rev1")) }
+        coVerify(exactly = 1) { mockRepo.deleteLocalTeamRecords(listOf("team3")) }
+        coVerify(exactly = 0) { retryQueue.queueFailedOperation(uploadType = "MyTeam", error = any(), payload = any(), endpoint = "teams", httpMethod = "POST", dbId = "team2", modelClassName = "MyTeam") }
     }
 
     @Test
     fun `uploadTeams handles bulk network failure`() = testScope.runTest {
-        coEvery { teamsUploadRepository.uploadTeams() } returns Unit
-        teamsUploadRepository.uploadTeams()
+        val mockRepo = io.mockk.mockk<TeamsSyncRepository>(relaxed = true)
+        every { teamsSyncRepository.get() } returns mockRepo
+
+        val mockTeam = TeamUploadData("team1", JsonObject(), false, null)
+        coEvery { mockRepo.getTeamsForUpload() } returns listOf(mockTeam)
+
+        val errorBody = okhttp3.ResponseBody.create(null, "Error")
+        coEvery { uploadRepository.postUploadArray(any(), any()) } returns retrofit2.Response.error(500, errorBody)
+        coEvery { retryQueue.queueFailedOperation(any(), any(), any(), any(), any(), any(), any()) } returns Unit
+
+        uploadManager.uploadTeams()
         advanceUntilIdle()
-        coVerify { teamsUploadRepository.uploadTeams() }
+
+        coVerify(exactly = 1) { uploadRepository.postUploadArray("http://mock.url/teams/_bulk_docs", any()) }
+        coVerify(exactly = 1) { retryQueue.queueFailedOperation(uploadType = "MyTeam", error = any(), payload = any(), endpoint = "teams", httpMethod = "POST", dbId = "team1", modelClassName = "MyTeam") }
     }
 
     @Test
@@ -85,13 +117,11 @@ class UploadManagerTest {
         coEvery { uploadRepository.postUploadArray(any(), any()) } throws java.io.IOException("Network down")
         coEvery { retryQueue.queueFailedOperation(any(), any(), any(), any(), any(), any(), any()) } returns Unit
 
-        teamsUploadRepository.uploadTeams()
+        uploadManager.uploadTeams()
         advanceUntilIdle()
 
-        coVerify { teamsUploadRepository.uploadTeams() }
-        // coVerify(exactly = 1) { uploadRepository.postUploadArray("http://mock.url/teams/_bulk_docs", any()) }
-        coVerify { teamsUploadRepository.uploadTeams() }
-        // coVerify(exactly = 1) { retryQueue.queueFailedOperation(uploadType = "MyTeam", error = any(), payload = any(), endpoint = "teams", httpMethod = "POST", dbId = "team1", modelClassName = "MyTeam") }
+        coVerify(exactly = 1) { uploadRepository.postUploadArray("http://mock.url/teams/_bulk_docs", any()) }
+        coVerify(exactly = 1) { retryQueue.queueFailedOperation(uploadType = "MyTeam", error = any(), payload = any(), endpoint = "teams", httpMethod = "POST", dbId = "team1", modelClassName = "MyTeam") }
     }
 
     private lateinit var uploadManager: UploadManager
@@ -100,7 +130,6 @@ class UploadManagerTest {
     private val sharedPrefManager: SharedPrefManager = mockk(relaxed = true)
     private val gson: Gson = mockk(relaxed = true)
     private val uploadCoordinator: UploadCoordinator = mockk(relaxed = true)
-    private val teamsUploadRepository: TeamsUploadRepository = mockk(relaxed = true)
     private val uploadRepository: UploadRepository = mockk(relaxed = true)
     private val retryQueue: RetryQueue = mockk(relaxed = true)
     private val personalsRepository: PersonalsRepository = mockk(relaxed = true)
@@ -156,8 +185,7 @@ class UploadManagerTest {
                 testScope,
                 photoUploader,
                 achievementUploader,
-                TestTimeProvider(),
-                teamsUploadRepository
+                TestTimeProvider()
             )
         )
     }
@@ -316,23 +344,25 @@ class UploadManagerTest {
 
     @Test
     fun `uploadResource returns early when no resources to upload`() = testScope.runTest {
-        coEvery { teamsUploadRepository.uploadResource(any()) } returns Unit
+        coEvery { uploadCoordinator.uploadRoom<MyLibrary>(any()) } returns UploadResult.Empty
         val listener = mockk<OnSuccessListener>(relaxed = true)
+
         uploadManager.uploadResource(listener)
         advanceUntilIdle()
-        coVerify { teamsUploadRepository.uploadResource(listener) }
+
+        coVerify { listener.onSuccess("No resources to upload") }
     }
 
     @Test
     fun `uploadResource notifies listener on failure`() = testScope.runTest {
         val errorMessage = "Test error"
-        coEvery { teamsUploadRepository.uploadResource(any()) } returns Unit
+        coEvery { uploadCoordinator.uploadRoom<MyLibrary>(any()) } throws Exception(errorMessage)
         val listener = mockk<OnSuccessListener>(relaxed = true)
 
         uploadManager.uploadResource(listener)
         advanceUntilIdle()
 
-        coVerify { teamsUploadRepository.uploadResource(listener) }
+        coVerify { listener.onSuccess("Resource upload failed: $errorMessage") }
     }
 
 }
